added tests for forecast
